requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
PyYAML==6.0.1
gunicorn==21.2.0
pytest==7.4.0
bandit==1.7.5
